requirements.txt
streamlit
opencv-python
numpy
pillow
rembg
scikit-image